Proof-read lib.rs documentation, updated User Guide links and references.
The lib.rs documentation can be finalized, after checking on the TODO around OpenGLWindow.
Also split NoWindow into its own file. Its documentation is also complete and finalized.
User Guide examples still need testing.
